Replaced the VideoStream constructor with CSICamera, added shebang line for python on jetson nano.
Replaced the VideoStream class from imutils with the CSICamera class from
Jetpack for jetson nano. This class provides an easy way to access the CSI camera interface on the jetson nano.
The default VideoStream class seems to be incompatible with the CSI Camera interface on the Jetson, but works fine when used with USB and built-in webcams ( My pc's webcam for example).

Also added a shebang line for indicating the location of the python3 executable on the Jetson Nano.
This path may change depending on your installation location. So please verify when using it on your own device.