Changed stream resolution. Reduced the frame counter threshold to increase sensitivity.
Stream resolution changed to match available camera modes. The CSI Camera can be
operated only at specifc modes. Currently resolution is set at 1920*1080 at ~30fps.
VideoStream Import was removed as it is no longer required.

The frame counter threshold variable (EYE_AR_CONSEC_FRAMES) was reduced to 5 frames so as to increase the sensitivity.

One issue noted is the low framerate. This is possibly due to the fact that the CPU in the SBC is being used instead of the GPU. This ws confirmed by taking the GPU uitilization readings using jtop.
This issue needs to be fixed and GPU should be utilized for detection.